Rename file erc20_wrapper.py to erc20_token.py
To match the name of the underlying contract.
