+++ python-packages/contract_wrappers/src/zero_ex/contract_wrappers/__init__.py
@@ -296,5 +296,5 @@ HexBytes('0x...')
 """
 
 from .tx_params import TxParams
-from .erc20_wrapper import ERC20Token
+from .erc20_token import ERC20Token
 from .exchange_wrapper import Exchange
